+++ test/action_policy/policy/aliases_test.rb
@@ -8,16 +8,20 @@ class AliasesRuleTestPolicy
 
   default_rule :manage?
 
-  alias_rule :update?, :destroy?, to: :edit?
+  alias_rule :update?, :destroy?, :create?, to: :edit?
 
   def manage?; end
 
   def edit?; end
 
+  def index?; end
+
   class NoDefault < self
     default_rule nil
 
-    alias_rule :update?, to: :manage?
+    alias_rule :index?, :update?, to: :manage?
+
+    def create?; end
   end
 end
 
@@ -54,5 +58,7 @@ class TestPolicyAliasesRule < Minitest::Test
 
     assert_equal :edit?, policy.resolve_rule(:destroy?)
     assert_equal :manage?, policy.resolve_rule(:update?)
+    assert_equal :manage?, policy.resolve_rule(:index?)
+    assert_equal :create?, policy.resolve_rule(:create?)
   end
 end